Maintenance - Cleanup, Comments & Preparation
- Renderer : Condensed RenderAPI & Renderer.
- Framebuffer :  Added relativity Bounds for HoveredEntity.
- Camera : Updated view projection handling.
- Application : Added control over main Framebuffer.
- LayerStack : Removed cached index variable. Will assign Layer IDs as inserted/removed.
- Layer : Reworked to only render validated Entities using RenderComponent ID.
- Project : New serialized Scripting Core filepaths & Tag handling.
- AssetManager : Condensed AssetSerializer & AssetManager(Runtime(Binary), Editor(Text)). Added/Updated Script & Field handling.
- Scene : Seperated Entity/ScriptableEntity files. Removed Simulate State.
- Component : Renamed TagComponent->NameComponent. New TagComponent & RenderComponent(Layers).
- Scripting : Script Asset & Entity/Script Event Handling.
- Audio : Fixed updating of ListenerComponent position & velocity.
- Math : Created basic type wrapped for Renderer, Shader & Vertex Array/Buffer

- EditorStack : New. Inherits base ID map & defines String map for Layer names.
- EditorLayer : Added collider visualization options & now controls ImGUI.
- Scene/Asset Panel : Added/Updated visualization for Tag/Render(Layer)/Script Component.

+++ game-engine/src/Platform/OpenAL/Assets/OpenALAudio.h
@@ -16,7 +16,7 @@ namespace GE
 		const uint32_t& GetID() const override { return m_Config.BufferIDs.front(); }
 		const std::vector<uint32_t>& GetAllID() const { return m_Config.BufferIDs; }
 		const Config& GetConfig() const override { return m_Config; }
-		const uint32_t GetBufferCount() const override;
+		const uint64_t GetBufferCount() const override;
 
 		const float GetDurationInSeconds() override;
 		const std::vector<uint32_t>& GenerateBuffers(const uint32_t& count = 1) override;
